requirements.txt
requests
sscws
python-dotenv==1.0.1